fix: resolve all clippy warnings for CI (-Dwarnings)
- Remove dead code (insert_audit_in_tx, unused admin helper)
- Fix collapsible_if, redundant closures, needless returns
- Fix octal escape in test (use \0 for null byte)
- Remove unused imports (sha2, test helpers)
- Use derive(Default) instead of manual impl
- Use repeat_n() instead of repeat().take()
- Apply cargo clippy --fix for trivial lints

+++ crates/dbward-server/src/background.rs
@@ -254,19 +254,19 @@ pub(crate) async fn run_record_purge_once(state: &AppState, retention: &Retentio
 
     // Revoked tokens
     match state.token_repo.purge_revoked(&request_cutoff) {
-        Ok(n) => { if n > 0 { result.processed += n as u32; info!(task = "record_purge", count = n, "purged revoked tokens"); } }
+        Ok(n) => { if n > 0 { result.processed += n; info!(task = "record_purge", count = n, "purged revoked tokens"); } }
         Err(e) => { result.failed += 1; error!(task = "record_purge", error = %e, "purge_revoked failed"); }
     }
 
     // Old audit events
     match state.audit_repo.purge_old(&audit_cutoff) {
-        Ok(n) => { if n > 0 { result.processed += n as u32; info!(task = "record_purge", count = n, "purged old audit events"); } }
+        Ok(n) => { if n > 0 { result.processed += n; info!(task = "record_purge", count = n, "purged old audit events"); } }
         Err(e) => { result.failed += 1; error!(task = "record_purge", error = %e, "purge_old audit failed"); }
     }
 
     // Old requests
     match state.request_repo.purge_old_requests(&request_cutoff) {
-        Ok(n) => { if n > 0 { result.processed += n as u32; info!(task = "record_purge", count = n, "purged old requests"); } }
+        Ok(n) => { if n > 0 { result.processed += n; info!(task = "record_purge", count = n, "purged old requests"); } }
         Err(e) => { result.failed += 1; error!(task = "record_purge", error = %e, "purge_old_requests failed"); }
     }
